Add files via upload
In this assignment, we have explored the dataset at the beginning of model development. We find out numerical and categorical predictive variables, fix missing values, find out rare variables, plot their distribution to see if there is any skewness in the data. After the data analysis, we separate the data into train and test set and fit the data by applying logistic regression function from scikit-learn. We evaluate the model performance on testset. It provides expected accuracy and roc-auc score.